fix: prevent NPE on player death from spear item
Combat system may assign `null` as instigator to some (Before)DestroyEvents which causes an NPE in the player death system.

This fixes the crash with a `TODO` item to find and fix the root cause. See https://github.com/Terasology/CombatSystem/pull/92 for first investigation results.

In addition, this commit introduces an enum for the statistics type to update to avoid any bugs resulting from using plain string "constants" (we did not even use constants ...).